+++ packages/manager/apps/sms/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.6.15](https://github.com/ovh/manager/compare/@ovh-ux/manager-sms-app@7.6.14...@ovh-ux/manager-sms-app@7.6.15) (2023-01-10)
+
+**Note:** Version bump only for package @ovh-ux/manager-sms-app
+
+
+
+
+
 ## [7.6.14](https://github.com/ovh/manager/compare/@ovh-ux/manager-sms-app@7.6.13...@ovh-ux/manager-sms-app@7.6.14) (2022-12-20)