+++ packages/manager/apps/telecom-task/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.8.2](https://github.com/ovh/manager/compare/@ovh-ux/manager-telecom-task-app@6.8.1...@ovh-ux/manager-telecom-task-app@6.8.2) (2023-01-10)
+
+**Note:** Version bump only for package @ovh-ux/manager-telecom-task-app
+
+
+
+
+
 ## [6.8.1](https://github.com/ovh/manager/compare/@ovh-ux/manager-telecom-task-app@6.8.0...@ovh-ux/manager-telecom-task-app@6.8.1) (2022-12-20)