+++ packages/manager/modules/mfa-enrollment/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.4.0](https://github.com/ovh/manager/compare/@ovh-ux/mfa-enrollment@2.3.3...@ovh-ux/mfa-enrollment@2.4.0) (2023-01-10)
+
+
+### Features
+
+* remove mfa-enrollment module ([4280f97](https://github.com/ovh/manager/commit/4280f978c4937735773eae0787dc14e1b00ec4ab))
+
+
+
+
+
 ## [2.3.3](https://github.com/ovh/manager/compare/@ovh-ux/mfa-enrollment@2.3.2...@ovh-ux/mfa-enrollment@2.3.3) (2022-11-29)
 
 **Note:** Version bump only for package @ovh-ux/mfa-enrollment